less casting
removing incessant u64 to f64 casting around germ genotyper

+++ src/kplib/annotator.rs
@@ -133,8 +133,8 @@ fn haploid(
 ) -> GenotypeAnno {
     if paths.is_empty() {
         let handle = match coverage {
-            0 => (".", metrics::GTstate::Non, 0.0, true),
-            _ => ("0", metrics::GTstate::Ref, 0.0, true),
+            0 => (".", metrics::GTstate::Non, 0, true),
+            _ => ("0", metrics::GTstate::Ref, 0, true),
         };
         return finalize_annotation(handle, paths, coverage, neigh_group, sample_idx, *var_idx);
     }
@@ -144,23 +144,23 @@ fn haploid(
         true => (
             "1",
             metrics::GTstate::Hom,
-            path1.meta.coverage[sample_idx] as f64,
+            path1.meta.coverage[sample_idx],
             true,
         ),
-        false if coverage != 0 => ("0", metrics::GTstate::Ref, 0.0, true),
-        false => (".", metrics::GTstate::Non, 0.0, true),
+        false if coverage != 0 => ("0", metrics::GTstate::Ref, 0, true),
+        false => (".", metrics::GTstate::Non, 0, true),
     };
     finalize_annotation(handle, paths, coverage, neigh_group, sample_idx, *var_idx)
 }
 
 /// GT str, GTstate, alt_cov, is_fulltarget
-type HandleReturn<'a> = (&'a str, metrics::GTstate, f64, bool);
+type HandleReturn<'a> = (&'a str, metrics::GTstate, u64, bool);
 
 fn handle_diploid_no_paths<'a>(coverage: u64) -> HandleReturn<'a> {
     if coverage != 0 {
-        ("0|0", metrics::GTstate::Ref, 0.0, true)
+        ("0|0", metrics::GTstate::Ref, 0, true)
     } else {
-        ("./.", metrics::GTstate::Non, 0.0, true)
+        ("./.", metrics::GTstate::Non, 0, true)
     }
 }
 
@@ -171,10 +171,10 @@ fn handle_diploid_single_path<'a>(
     sample_idx: usize,
 ) -> HandleReturn<'a> {
     if !path.path.contains(var_idx) {
-        ("0|0", metrics::GTstate::Ref, 0.0, true)
+        ("0|0", metrics::GTstate::Ref, 0, true)
     } else {
-        let alt_cov = path.meta.coverage[sample_idx] as f64;
-        let ref_cov = (coverage as f64) - alt_cov;
+        let alt_cov = path.meta.coverage[sample_idx];
+        let ref_cov = coverage - alt_cov;
         let (genotype, state) = match metrics::genotyper(ref_cov, alt_cov) {
             metrics::GTstate::Ref | metrics::GTstate::Het => {
                 let gt = match path.meta.hp[sample_idx] {
@@ -202,23 +202,23 @@ fn handle_diploid_two_paths<'a>(
         (true, true) => (
             "1|1",
             metrics::GTstate::Hom,
-            (path1.meta.coverage[sample_idx] + path2.meta.coverage[sample_idx]) as f64,
+            (path1.meta.coverage[sample_idx] + path2.meta.coverage[sample_idx]),
             path1.full_target || path2.full_target,
         ),
         (true, false) => (
             "1|0",
             metrics::GTstate::Het,
-            path1.meta.coverage[sample_idx] as f64,
+            path1.meta.coverage[sample_idx],
             path1.full_target,
         ),
         (false, true) => (
             "0|1",
             metrics::GTstate::Het,
-            path2.meta.coverage[sample_idx] as f64,
+            path2.meta.coverage[sample_idx],
             path2.full_target,
         ),
-        (false, false) if coverage != 0 => ("0|0", metrics::GTstate::Ref, 0.0, true),
-        (false, false) => ("./.", metrics::GTstate::Non, 0.0, true),
+        (false, false) if coverage != 0 => ("0|0", metrics::GTstate::Ref, 0, true),
+        (false, false) => ("./.", metrics::GTstate::Non, 0, true),
     }
 }
 
@@ -231,7 +231,7 @@ fn finalize_annotation(
     var_idx: NodeIndex,
 ) -> GenotypeAnno {
     let (gt_str, gt_path, alt_cov, full_target) = handle;
-    let ref_cov = coverage as f64 - alt_cov;
+    let ref_cov = coverage - alt_cov;
 
     let gt_obs = metrics::genotyper(ref_cov, alt_cov);
 
@@ -268,7 +268,7 @@ fn finalize_annotation(
         if sq < 5.0 {
             filt |= FiltFlags::LOWSQ;
         }
-        if alt_cov < 5.0 {
+        if alt_cov < 5 {
             filt |= FiltFlags::LOWALT;
         }
     }
